+++ history.md
@@ -15,6 +15,7 @@
 
 ### 4.2.19
 - 简化SelectValue设置项，删除dict相关参数直接使用默认值，特殊项目可单独生成SelectValue类适配，减少内存占用
+- 拆分CascadeValue和SelectValue
 
 ### 4.2.18
 - SearchData的getData默认深拷贝，避免在获取数据后改变数据导致问题